feat: implementar sumatoria diaria de asistencias e inasistencias en base a pesos de horarios semanales

+++ backend/src/main/java/com/school/attendance/dto/report/DailyCourseSummaryDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class DailyCourseSummaryDTO {
-    private int presentCount;
-    private int absentCount;
+    private double presentCount;
+    private double absentCount;
     private int totalStudents;
 }
